+++ src/core/vector/qgsvectorlayerfeaturecounter.cpp
@@ -101,12 +101,12 @@ bool QgsVectorLayerFeatureCounter::run()
   return true;
 }
 
-QHash<QString, long> QgsVectorLayerFeatureCounter::symbolFeatureCountMap() const
+QHash<QString, long long> QgsVectorLayerFeatureCounter::symbolFeatureCountMap() const
 {
   return mSymbolFeatureCountMap;
 }
 
-long QgsVectorLayerFeatureCounter::featureCount( const QString &legendKey ) const
+long long QgsVectorLayerFeatureCounter::featureCount( const QString &legendKey ) const
 {
   return mSymbolFeatureCountMap.value( legendKey, -1 );
 }
